+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ### next
 - gray panicking on out of range level has been changed to a silent clamping
+- skin now deserializable with serde
+- several utilities to parse colors and style elements
 
 <a name="v0.23.2"></a>
 ### v0.23.2 - 2023-08-02